Add telephony binding event support and UI call status notification
Add supports for incoming call, dialing call, and terminated call events
in the telephony binding. The phone UI is enhanced to make use of these
telephony binding events to display a notifications of phone call
status. These include generate a ring tone and displaying incoming phone
number information, outgoing phone number being dialed, and halt of the
ring tone and clearing of the notification space when a call is
terminated.

AGL-Bug: SPEC-598
Change-Id: Ied610b70c2e6edb1f631decd417cdbd39746a558
Signed-off-by: Matt Porter <mporter@konsulko.com>