Refactor telephony binding voicecall object management
Refactor the voicecall object lifecycle management in the
telephony binding using g_object-derived handles. Instead
of passing around and managing raw strings as a handle to
a voicecall, we use the voicecall object itself returned
from the new dbus proxy init. This simplifies memory
management of these objects that are created and freed
each time as calls are dialed/answered and terminated. It
then simplifies the forthcoming call answering support.

AGL-Bug: SPEC-600
Change-Id: I89efed1eb7927f4ffe2d1eaed295e0aa914efdcc
Signed-off-by: Matt Porter <mporter@konsulko.com>