Add support for answering calls to the telephony binding and UI
Adds an answer verb and associated incoming call lifecycle
management to the telephony binding. Leverages the binding
answer support to activate the call answer button when the
binding reports that we are in an incoming call state. Once
the incoming call is answered, the ring tone is deactivated.

AGL-Bug: SPEC-600
Change-Id: I4f2db2d248cfd0f4945fc17d450e9b691061cc9e
Signed-off-by: Matt Porter <mporter@konsulko.com>