Add incoming call rejection to telephony binding and phone UI
Adds support for disconnection of an alerting incoming call to
the telephony binding via the hangup verb. This is support in
the phone UI by an additional button being loaded in the
incoming call state allowing the user to decline to answer the
call or use the already present answer call button to accept
it.

AGL-Bug: SPEC-601
Change-Id: I56ad0b813c0f1f24688709d180c3c061d7511651
Signed-off-by: Matt Porter <mporter@konsulko.com>